presence_sim_addon/requirements.txt
flask==3.0.3
requests==2.32.3
apscheduler==3.10.4
python-dateutil==2.9.0.post0
